% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{a_bic}
\alias{a_bic}
\title{Calculating AIC or BIC value for a given beta}
\usage{
a_bic(beta, ncat, dat, label, y, cen, type, nlambda, eps, useBIC = F)
}
\arguments{
\item{beta}{An object of class "param".}

\item{ncat}{A vector specifying the number of levels for each risk factor.}

\item{dat}{The dataset.}

\item{label}{The combinations of risk factors.}

\item{y}{A vector representing the time to event or binary outcome.}

\item{cen}{A vector indicating the censoring status, or the binary outcome.}

\item{type}{The outcome type, either survival outcome or binary outcome.}

\item{nlambda}{The size of the search space for the tuning parameter.}

\item{eps}{The tolerance level.}

\item{useBIC}{A boolean indicating whether BIC is used.}
}
\value{
The AIC or BIC value. If beta contains a single lambda value, the function returns a single AIC or BIC value. If beta contains a vector of lambda values, the function returns a vector of AIC or BIC values.
}
\description{
This function calculates the Akaike Information Criterion (AIC) or Bayesian Information Criterion (BIC) value for a given beta.
}
