% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{as_inequalities}
\alias{as_inequalities}
\title{Generating the partial ordering matrix}
\usage{
as_inequalities(Hasse, cellName)
}
\arguments{
\item{Hasse}{The partial ordering relations expressed by a directed graph.}

\item{cellName}{The labels of the nodes in the Hasse diagram.}
}
\value{
A matrix expressing the partial ordering constraints on coefficients.
}
\description{
The function performs a transformation on the Hasse diagram to generate a matrix that expresses the partial ordering constraints on coefficients. The resulting matrix represents the constraints imposed by the partial ordering relations.
}
