% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bric_bc_os.R
\docType{data}
\name{bric_bc_os}
\alias{bric_bc_os}
\title{Fetching an example data from METABRIC}
\format{
An object of class \code{data.frame} with 1238 rows and 9 columns.
}
\source{
For more information about the METABRIC study and the dataset, visit: \href{https://www.cbioportal.org/study/summary?id=brca_metabric}{METABRIC Study Summary}
}
\usage{
data("bric_bc_os")
}
\description{
This function fetches an example dataset from the METABRIC (Molecular Taxonomy of Breast Cancer International Consortium) study.
}
\examples{
data("bric_bc_os")
}
\keyword{datasets}
