% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{cffStg}
\alias{cffStg}
\title{Getting the constrained estimates for cancer stages}
\usage{
cffStg(
  coeff0,
  theta0,
  cen,
  y,
  stage,
  Cov,
  withCov = FALSE,
  maxiter = 20,
  eps = 10^-4,
  getLikelihood = FALSE,
  fullModel = FALSE,
  fullZ = NULL,
  type = "surv",
  seed = 0
)
}
\arguments{
\item{coeff0}{A vector of initial coefficients for nodes.}

\item{theta0}{A vector of initial coefficients for covariates that need adjusting.}

\item{cen}{A vector of censoring statuses, or the binary outcome.}

\item{y}{A vector of outcomes.}

\item{stage}{A vector of stages after initial classification.}

\item{Cov}{A matrix of covariates.}

\item{withCov}{Whether any covariates need adjusting.}

\item{maxiter}{The maximum iteration in each approximation step.}

\item{eps}{A numeric value denoting the accuracy level.}

\item{getLikelihood}{Whether it outputs the value of the negative likelihood.}

\item{fullModel}{Whether it is a full model, representing the model before pruning to the current model.}

\item{fullZ}{A matrix of stages for the full model.}

\item{type}{Outcome type, either survival outcome or binary outcome.}

\item{seed}{Seed used for generating the simulation dataset.}
}
\value{
The estimates for cancer stages given the partial ordering constraints.
}
\description{
Finds the estimates for cancer stages given the partial ordering constraints.
}
