% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CnearPD}
\alias{CnearPD}
\title{Compute Nearest Positive Definite Matrix}
\usage{
CnearPD(mat)
}
\arguments{
\item{mat}{An approximately positive definite matrix.}
}
\value{
The nearest positive definite matrix.
}
\description{
This function computes the nearest positive definite matrix to an approximately positive definite matrix.
}
