% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countDistinct}
\alias{countDistinct}
\title{Count Non-Zero Distinct Values in Vector}
\usage{
countDistinct(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
The number of non-zero distinct values in the vector.
}
\description{
This function calculates the number of non-zero distinct values in a given vector.
}
