% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{createRiskGroups}
\alias{createRiskGroups}
\title{Generating Risk Categories as Separate Variables}
\usage{
createRiskGroups(dat, numRiskFs)
}
\arguments{
\item{dat}{The dataset that contains a variable named \code{variable} with values such as \code{a1b2} indicating the first level for the risk factor \code{a} and the second level for the risk factor \code{b}.}

\item{numRiskFs}{The total number of risk factors.}
}
\value{
The dataset with separate variables indicating different ordinal risk factors.
}
\description{
This function constructs a dataset with separate variables indicating different ordinal risk factors based on a given dataset with a variable named \code{variable} that represents risk categories.
}
\keyword{dataset,}
\keyword{factors,}
\keyword{ordinal}
\keyword{risk}
\keyword{variables}
