% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumSum}
\alias{cumSum}
\title{Calculate Cumulative Sums of Vector Elements}
\usage{
cumSum(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
A vector containing the cumulative sums of the input vector.
}
\description{
This function calculates the cumulative sums of elements in a given vector.
}
