% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{dataset.gen}
\alias{dataset.gen}
\title{Simulating a dataset}
\usage{
dataset.gen(
  beta.mat,
  p.mat,
  n = 800,
  seed = 2021,
  cen.prob = 0.8,
  label,
  ordered = TRUE,
  type = "surv",
  firstCon = FALSE
)
}
\arguments{
\item{beta.mat}{The betas or specified coefficients in matrix form for all nodes.}

\item{p.mat}{The probabilities of a patient being sampled from nodes in matrix form.}

\item{n}{The total number of patients.}

\item{seed}{The seed used for reproducibility.}

\item{cen.prob}{The proportion of censored patients.}

\item{label}{The labels of nodes.}

\item{ordered}{Whether the dataset should be sorted by survival time in ascending order.}

\item{type}{The outcome type, which can be either survival outcome or binary outcome.}

\item{firstCon}{Whether the first variable is continuous.}
}
\value{
A data frame representing one dataset.
}
\description{
The function generates a data frame representing one dataset based on the provided inputs.
}
