% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derivatives}
\alias{derivatives}
\title{Calculate First and Second Derivatives of Log (Partial) Likelihood}
\usage{
derivatives(Z, cen, beta, y, cov, theta, seed, withCov = FALSE, type = "surv")
}
\arguments{
\item{Z}{A matrix of nodes.}

\item{cen}{A vector of censoring statuses or binary outcomes.}

\item{beta}{A vector of node coefficients.}

\item{y}{A vector of outcomes.}

\item{cov}{Covariates.}

\item{theta}{Covariate coefficients.}

\item{seed}{The seed used for generating the simulation dataset.}

\item{withCov}{Whether any covariates need adjusting.}

\item{type}{A string indicating the type of outcome, currently either "survival" or "binary".}
}
\value{
A list of derivatives.
}
\description{
This function calculates the first and second derivatives of the log (partial) likelihood based on the given parameters.
}
