% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{dM}
\alias{dM}
\title{Generating a dummy matrix for a categorical variable}
\usage{
dM(X, ref = TRUE)
}
\arguments{
\item{X}{The categorical variable provided as a vector.}

\item{ref}{Whether a reference level is needed. If set to \code{TRUE}, the first level is chosen as the reference level.}
}
\value{
A dummy matrix representing the categorical variable.
}
\description{
The function generates a dummy matrix where each column represents a category or level of the categorical variable. The matrix contains binary values (0 or 1) indicating whether each observation falls into a particular category. If a reference level is specified, the dummy matrix will include one less column compared to the total number of levels.
}
