% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ED}
\alias{ED}
\title{Call R Function to Compute Eigenvalues and Eigenvectors}
\usage{
ED(mat)
}
\arguments{
\item{mat}{An approximately positive definite matrix.}
}
\value{
A list containing the eigenvalues and eigenvectors.
}
\description{
This C++ function calls an R function to compute the eigenvalues and eigenvectors of a numeric (double, integer, logical) or complex matrix.
}
