% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{edgesHasse}
\alias{edgesHasse}
\title{Generating the edges}
\usage{
edgesHasse(ncat, e, vs = NULL)
}
\arguments{
\item{ncat}{A vector specifying the number of categories in each ordinal variable or risk factor with a total ordering.}

\item{e}{The current edges in the Hasse diagram, starting from an empty vector.}

\item{vs}{The current vertex.  When it is null, it starts from (1, 1, ..., 1) depending on the number of variables.}
}
\value{
A vector of edges representing the connections between vertices in the Hasse diagram.
}
\description{
The function generates edges in the Hasse diagram based on the given inputs. It then returns a vector of edges representing the ordered relationships between vertices in the Hasse diagram.
}
\examples{
edgesHasse(ncat = c(2, 3, 2), e = vector())

}
\seealso{
\code{\link{nextEdges}}
}
