% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{generatePalette}
\alias{generatePalette}
\title{Generating a list regarding color assignments and true stages}
\usage{
generatePalette(beta.mat, label, c = "Viridis")
}
\arguments{
\item{beta.mat}{The betas in matrix form.}

\item{label}{The labels of nodes.}

\item{c}{The basic color palette.}
}
\value{
A list containing colors for nodes, unique coefficients, the color palette, and the true stage for nodes.
}
\description{
The function generates a list that contains the following elements: colors for nodes, unique coefficients, the color palette, and the true stage for nodes.
}
