% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMin}
\alias{getMin}
\title{Calculating the minimum of a vector}
\usage{
getMin(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
The minimum value in the vector.
}
\description{
This function calculates the minimum value of a given vector.
}
