% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{getsEst}
\alias{getsEst}
\title{Calculating the estimated coefficients for stages and the corresponding value for the chosen criterion}
\usage{
getsEst(
  Data,
  cen,
  y,
  curR,
  Cov,
  withCov,
  type,
  useIbs = F,
  useAIC = F,
  seed = 0,
  fullModel = FALSE,
  fullZ = NULL,
  GIC = 2
)
}
\arguments{
\item{Data}{The given dataset.}

\item{cen}{A vector of censoring statuses, or the binary outcome.}

\item{y}{A vector of outcomes.}

\item{curR}{The current staging result.}

\item{Cov}{A matrix of covariates.}

\item{withCov}{A boolean indicating whether any covariates need adjusting.}

\item{type}{The type of outcome, either survival outcome or binary outcome.}

\item{useIbs}{A boolean indicating whether the (integrated) Brier score is applied.}

\item{useAIC}{A boolean indicating whether AIC (Akaike Information Criterion) is used}

\item{seed}{The seed used for generating the simulation dataset.}

\item{fullModel}{A boolean indicating whether it is a full model, representing the model before pruning to the current model.}

\item{fullZ}{A matrix of stages for the full model.}

\item{GIC}{The penalty term used in the criterion. The default value is 2 when AIC is used.}
}
\value{
The estimated coefficients for stages and the corresponding value for the chosen criterion.
}
\description{
This function calculates the estimated coefficients for stages and the corresponding value for the chosen criterion. It takes a dataset, censoring statuses, outcomes, current staging result, covariates, and other parameters as inputs.
}
