% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gs}
\alias{gs}
\title{Create Equally-Spaced Logarithmic Sequence}
\usage{
gs(a, b, c = 30L)
}
\arguments{
\item{a}{A numeric value representing the starting value.}

\item{b}{A numeric value representing the ending value.}

\item{c}{The length of the sequence.}
}
\value{
An equally-spaced sequence in the log scale (a geometric series).
}
\description{
This function creates an equally-spaced sequence in the log scale, also known as a geometric series, between a starting value and an ending value.
}
