% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IRLSO}
\alias{IRLSO}
\title{Perform One Iteration of Iterative Re-weighted Least Squares}
\usage{
IRLSO(
  coeff0,
  theta0,
  cen,
  y,
  Z,
  cov,
  Cnstrn,
  coeffLambda,
  seed = 0L,
  withCov = FALSE,
  type = "surv"
)
}
\arguments{
\item{coeff0}{Current coefficients for nodes.}

\item{theta0}{Current coefficients for covariates that need adjusting.}

\item{cen}{A vector of censoring statuses or binary outcomes}

\item{y}{A vector of outcomes.}

\item{Z}{A matrix of nodes.}

\item{cov}{A matrix of covariates.}

\item{Cnstrn}{A matrix of partial ordering constraints.}

\item{coeffLambda}{A numeric value of the tuning parameter lambda.}

\item{seed}{The seed used for generating the simulation dataset.}

\item{withCov}{Whether any covariates need adjusting.}

\item{type}{A string indicating the type of outcome, either "survival" or "binary".}
}
\value{
A vector of coefficients at the next iteration.
}
\description{
This function performs one iteration of the Iterative Re-weighted Least Squares (IRLS) algorithm based on the given parameters.
}
