% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IRLSP}
\alias{IRLSP}
\title{Perform One Iteration of Iterative Re-weighted Least Squares in the Pruning Step}
\usage{
IRLSP(
  coeff0,
  theta0,
  cen,
  y,
  Z,
  cov,
  LCnstrn,
  qCnstrn,
  mCnstrn,
  minCoeff,
  maxCoeff,
  Lambda,
  withCov = FALSE,
  seed = 0L,
  type = "surv"
)
}
\arguments{
\item{coeff0}{Current coefficients for nodes.}

\item{theta0}{Current coefficients for covariates that need adjusting.}

\item{cen}{A vector of censoring statuses or binary outcomes.}

\item{y}{A vector of outcomes.}

\item{Z}{A matrix of nodes.}

\item{cov}{A matrix of covariates.}

\item{LCnstrn}{A matrix of linear partial ordering constraints.}

\item{qCnstrn}{A matrix of quadratic constraints.}

\item{mCnstrn}{An integer denoting how many more constraints are needed.}

\item{minCoeff}{The lower bound for all the coefficients.}

\item{maxCoeff}{The upper bound for all the coefficients.}

\item{Lambda}{A numeric value of the tuning parameter lambda to enforce the quadratic constraints.}

\item{withCov}{Whether any covariates need adjusting.}

\item{seed}{The seed used for generating the simulation dataset.}

\item{type}{A string indicating the type of outcome, either "survival" or "binary" outcome.}
}
\value{
A vector of coefficients at the next iteration.
}
\description{
This function performs one iteration of the Iterative Re-weighted Least Squares (IRLS) algorithm in the pruning step based on the given parameters.
}
