% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iter_pen}
\alias{iter_pen}
\title{Perform Iterative Solution of Group Fused Lasso Penalty}
\usage{
iter_pen(
  Dmat0,
  dvec0,
  Apo,
  lambda,
  w,
  nb,
  inibeta,
  maxiter,
  eps,
  trace = FALSE
)
}
\arguments{
\item{Dmat0}{The original quadratic term matrix.}

\item{dvec0}{The original linear term vector.}

\item{Apo}{The partial order matrix.}

\item{lambda}{The tuning parameter.}

\item{w}{The weights.}

\item{nb}{The number of boundaries.}

\item{inibeta}{The initial beta value for warm start.}

\item{maxiter}{The maximum number of iterations.}

\item{eps}{The tolerance number.}

\item{trace}{Whether to trace each iteration output.}
}
\value{
The solution beta.
}
\description{
This function performs the iterative solution of the group fused lasso penalty based on the given parameters.
}
