% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{lasso.tree.bic}
\alias{lasso.tree.bic}
\title{Fitting the Lasso Tree Method with Parameter Tuning}
\usage{
lasso.tree.bic(
  ncat,
  y,
  cen,
  Apo,
  label,
  dat,
  inibeta = NULL,
  trace = F,
  maxiter = 200,
  eps = 1e-04,
  type = "surv",
  giveNB = F,
  fullA = F,
  useBIC = F,
  N_lasso = NULL,
  ...
)
}
\arguments{
\item{ncat}{A vector specifying the number of levels for each risk factor.}

\item{y}{The time to failure, or the binary outcome.}

\item{cen}{An indicator for censoring, or the binary outcome.}

\item{Apo}{Theee partial order constraints.}

\item{label}{The combinations of risk factors.}

\item{dat}{The dataset.}

\item{inibeta}{A vector of initial values for the beta coefficients.}

\item{trace}{A boolean variable indicating whether to show the calculation process.}

\item{maxiter}{The maximum number of iterations.}

\item{eps}{The tolerance level.}

\item{type}{The outcome type, either survival outcome or binary outcome.}

\item{giveNB}{A boolean indicating whether the number of boundaries is given.}

\item{fullA}{A boolean indicating whether the second derivatives are fully derived.}

\item{useBIC}{A boolean indicating whether BIC is used.}

\item{N_lasso}{The size of the search space for tuning parameter}
}
\value{
A list that includes the best solution of beta according to the AIC or BIC, a vector of AIC or BIC values corresponding to the input lambda, and a \code{param} object called \code{beta.seq} that includes all the coefficients.
}
\description{
This function is the core function for fitting the Lasso Tree method with parameter tuning.
}
