% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{lasso.tree}
\alias{lasso.tree}
\title{Fitting the lasso tree method}
\usage{
lasso.tree(
  ncat,
  cen,
  y,
  lambda,
  dat,
  label,
  Apo,
  inibeta = NULL,
  trace = F,
  maxiter = 200,
  eps = 1e-04,
  type = "surv",
  giveNB = F,
  fullA = F
)
}
\arguments{
\item{ncat}{A vector specifying the number of levels for each risk factor.}

\item{cen}{An indicator for censoring, or the binary outcome.}

\item{y}{The time to failure, or the binary outcome.}

\item{lambda}{The tuning parameter of the fused group lasso penalty. It can be a single value or a vector.}

\item{dat}{The dataset.}

\item{label}{Combinations of risk factors.}

\item{Apo}{Partial order constraints.}

\item{inibeta}{A vector of initial guesses for the beta coefficients.}

\item{trace}{A boolean variable indicating whether to show the process of calculation.}

\item{maxiter}{The maximum number of iterations.}

\item{eps}{The tolerance level.}

\item{type}{The outcome type, either survival outcome or binary outcome.}

\item{giveNB}{A boolean indicating whether the number of boundaries is given.}

\item{fullA}{A boolean indicating whether the second derivatives are fully derived.}
}
\value{
The constrained estimates using lasso tree method
}
\description{
This function is the core function for lasso tree method.
}
\seealso{
\code{\link{as.param}}
}
