% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lasso_tree_multi}
\alias{lasso_tree_multi}
\title{Calculate Best Coefficients using Lasso Tree Method with Parameter Tuning}
\usage{
lasso_tree_multi(
  X,
  indset,
  Apo,
  nb,
  w,
  cen,
  y,
  lambda,
  inibeta,
  maxiter,
  eps,
  trace,
  type = "surv",
  fullA = FALSE
)
}
\arguments{
\item{X}{The design matrix.}

\item{indset}{The risk set.}

\item{Apo}{The partial ordering matrix.}

\item{nb}{The number of boundaries.}

\item{w}{The adaptive parameter.}

\item{cen}{The vector of censoring indicators or binary outcomes.}

\item{y}{The vector of times to failure or binary outcomes.}

\item{lambda}{The tuning parameters.}

\item{inibeta}{The initial beta value for warm start.}

\item{maxiter}{The maximum number of iterations.}

\item{eps}{The tolerance number.}

\item{trace}{Whether to trace each iteration output.}

\item{type}{The type of outcome, either "surv" for survival outcome or "bin" for binary outcome.}

\item{fullA}{Boolean variable indicating whether to calculate all elements of the Hessian.}
}
\value{
A numeric matrix with the first column representing the values of lambda and the remaining columns representing the corresponding coefficients.
}
\description{
This function calculates the best coefficients using the lasso tree method with parameter tuning.
}
