% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lasso_tree_single}
\alias{lasso_tree_single}
\title{Calculate the best coefficients using lasso tree method without parameter tuning}
\usage{
lasso_tree_single(
  X,
  indset,
  Apo,
  nb,
  w,
  cen,
  y,
  lambda,
  inibeta,
  maxiter,
  eps,
  trace,
  fullA = FALSE,
  type = "surv"
)
}
\arguments{
\item{X}{The design matrix.}

\item{indset}{The risk set.}

\item{Apo}{The partial ordering matrix.}

\item{nb}{The number of boundaries.}

\item{w}{The adaptive parameter.}

\item{cen}{The vector of censoring indicators or binary outcomes.}

\item{y}{The vector of times to failure or binary outcomes.}

\item{lambda}{The tuning parameter.}

\item{inibeta}{The initial beta value for warm start.}

\item{maxiter}{The maximum number of iterations.}

\item{eps}{The tolerance number.}

\item{trace}{Whether to trace each iteration output.}

\item{fullA}{Boolean variable indicating whether to calculate all elements of the Hessian.}

\item{type}{The type of outcome, either "surv" for survival outcome or "bin" for binary outcome.}
}
\value{
A vector of the solution beta.
}
\description{
This function calculates the best coefficients using lasso tree method without parameter tuning
}
