% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{listDummies}
\alias{listDummies}
\title{Generating a list of dummy variables}
\usage{
listDummies(dat)
}
\arguments{
\item{dat}{The data matrix containing categorical variables.}
}
\value{
A list of dummy variables, where each element represents a categorical variable and is a binary vector indicating the presence or absence of each category or level.
}
\description{
The function generates a list where each element corresponds to a categorical variable in the data matrix. Each element of the list is a dummy variable, represented as a binary vector or column, indicating the presence or absence of each category or level of the categorical variable.
}
