% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logPL_C}
\alias{logPL_C}
\title{Calculate Negative Log Partial Likelihood}
\usage{
logPL_C(X, indset, cen, y, beta, type = "surv")
}
\arguments{
\item{X}{The design matrix.}

\item{indset}{The risk set matrix.}

\item{cen}{The vector of censoring indicators or binary outcomes.}

\item{y}{The vector of times to failure or binary outcomes.}

\item{beta}{The vector of coefficients.}

\item{type}{The type of outcome, either "surv" for survival outcome or "bin" for binary outcome.}
}
\value{
The negative log partial likelihood.
}
\description{
This function calculates the negative log partial likelihood based on the given parameters.
}
