% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Min}
\alias{Min}
\title{Calculate Minimum of Vector Elements}
\usage{
Min(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
The minimum value of the vector elements.
}
\description{
This function calculates the minimum value among the elements of a given vector.
}
