% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{nextEdges}
\alias{nextEdges}
\title{Getting all the edges and the subsequent vertices of the current vertex}
\usage{
nextEdges(v, ncat)
}
\arguments{
\item{v}{The starting vertex representing the current levels for all ordinal variables or risk factors.}

\item{ncat}{A vector specifying the number of categories in each variable or risk factor.}
}
\value{
A list containing a vector of all the edges of the current vertex and a matrix of the next vertices of the current vertex.
}
\description{
This function takes a starting vertex, which represents the current levels for all ordinal variables or risk factors, and a vector specifying the number of categories in each ordinal variable or risk factor. It then generates the edges for the current vertex, which represents the ordered relationships between vertices. The function returns a list that contains a vector of all edges of the current vertex and a matrix of the next vertices of the current vertex.
}
\examples{
nextEdges(v = c(1, 2, 1), ncat = c(2, 3, 2))

}
\seealso{
\code{\link{edgesHasse}}
}
