% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nNv}
\alias{nNv}
\title{Replace Negative Values with Zeros in a Vector}
\usage{
nNv(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
The non-negative version of the vector.
}
\description{
This function replaces all the negative values of a vector with zeros.
}
