% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{NTitle}
\alias{NTitle}
\title{Generating Node Labels for \code{visNetwork} without Stages Specified}
\usage{
NTitle(x, riskNames, riskFactors)
}
\arguments{
\item{x}{The dataset that contains a variable named \code{variable} with values such as \code{a1b2} indicating the first level for the risk factor \code{a} and the second level for the risk factor \code{b}.}

\item{riskNames}{The names of all risk factors.}

\item{riskFactors}{The list of names for all levels of all risk factors.}
}
\value{
The node labels with the name for the specified level for each risk factor.
}
\description{
This function generates node labels for \code{visNetwork} based on a given risk category, the names of all risk factors, and the names for all levels of each risk factor.
}
