% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{NTitleS}
\alias{NTitleS}
\title{Generating Node Labels for \code{visNetwork} with Stages Specified}
\usage{
NTitleS(x, riskNames, riskFactors, S)
}
\arguments{
\item{x}{The dataset that contains a variable named \code{variable} with values such as \code{a1b2} indicating the first level for the risk factor \code{a} and the second level for the risk factor \code{b}.}

\item{riskNames}{The names of all risk factors.}

\item{riskFactors}{The list of names for all levels of all risk factors.}

\item{S}{The vector indicating the corresponding stages.}
}
\value{
The node labels with the name for the specified level for each risk factor and the stage for each risk factor.
}
\description{
This function generates node labels for \code{visNetwork} based on a given risk category, the names of all risk factors, and the names for all levels of each risk factor. It also incorporates stages specified by the user.
}
