% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{operai}
\alias{operai}
\title{Fitting OPERA without pruning}
\usage{
operai(
  Data,
  TimeN,
  cenN,
  yN,
  covN,
  Z,
  Cnstrn,
  withCov = F,
  maxiter = 10,
  eps = 10^-4,
  AIC = TRUE,
  sPoint = 0.001,
  type = "surv",
  seed = 0,
  GIC = 2,
  N = NULL,
  preS = T,
  perc = 0.1,
  minObs = 30,
  ...
)
}
\arguments{
\item{Data}{The dataset}

\item{TimeN}{The variable name for survival times}

\item{cenN}{The variable name for censoring times}

\item{yN}{The variable name for binary outcomes}

\item{covN}{The variable names for covariates}

\item{Z}{A matrix representing the node to which each observation resides.}

\item{Cnstrn}{A matrix of constraints}

\item{withCov}{Whether any covariates need adjusting}

\item{maxiter}{Maximum number of iterations in each approximation step}

\item{eps}{A numeric accuracy level}

\item{AIC}{Whether AIC is used}

\item{sPoint}{A numeric value for an initial coefficient}

\item{type}{A string of outcome type, either survival ("surv") or binary ("bin") outcome}

\item{seed}{Seed used for generating the simulation dataset}

\item{GIC}{The penalty term used in the criterion to identify the optimal stage}

\item{N}{The size of the search space for tuning parameter}

\item{preS}{Whether the coefficients from the regression model are used as the initial coefficients}

\item{perc}{The minimum proportion of patients in each stage}

\item{minObs}{The minimum number of patients in each stage}
}
\value{
The initial staging result from OPERA without pruning
}
\description{
This function implements OPERA without pruning
}
