% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{operap}
\alias{operap}
\title{Fitting the pruning step for the OPERA algorithm}
\usage{
operap(
  cen,
  y,
  Z,
  Cnstrn,
  Cov,
  Data,
  withCov = F,
  maxiter = 10,
  eps = 10^-4,
  GIC = NULL,
  initial_result,
  useIbs = FALSE,
  useLRT = TRUE,
  useAIC = FALSE,
  coarse_pruning = FALSE,
  fine_pruning = TRUE,
  fine_pruning_quad = FALSE,
  prefix_stage = 5,
  threshold = 0.01,
  type = "surv",
  seed = 0,
  getsAll = F,
  perc = 0.1,
  minObs = 30,
  ...
)
}
\arguments{
\item{cen}{A vector specifying the censoring statuses, or the binary outcome.}

\item{y}{A vector representing the outcome variable.}

\item{Z}{A matrix of nodes.}

\item{Cnstrn}{A matrix of constraints.}

\item{Cov}{A matrix of covariates.}

\item{Data}{A given dataset.}

\item{withCov}{A boolean indicating whether any covariates need adjusting.}

\item{maxiter}{The maximum number of iterations in each approximation step.}

\item{eps}{The desired accuracy level.}

\item{GIC}{The penalty term used in the criterion. The default value is 2 when AIC is used.}

\item{initial_result}{The initial result from OPERA.}

\item{useIbs}{A boolean indicating whether the (integrated) Brier score is applied.}

\item{useLRT}{A boolean indicating whether the likelihood ratio test is applied.}

\item{useAIC}{A boolean indicating whether AIC (Akaike Information Criterion) is used.}

\item{coarse_pruning}{A boolean indicating whether coarse pruning is applied.}

\item{fine_pruning}{A boolean indicating whether fine pruning using exhaustive search is applied.}

\item{fine_pruning_quad}{A boolean indicating whether fine pruning using quadratic programming constraint is applied.}

\item{prefix_stage}{The number of stages needed.}

\item{threshold}{The type-I error rate for the likelihood ratio test.}

\item{type}{The outcome type, either survival outcome or binary outcome.}

\item{seed}{The seed used for generating the simulation dataset.}

\item{getsAll}{A boolean indicating whether pruning is performed until only two stages are left with the values for all the criteria}

\item{perc}{The minimum proportion of patients required in each stage}

\item{minObs}{The minimum number of patients required in each stage}
}
\value{
The cancer staging result after pruning.
}
\description{
This function performs the pruning step for the OPERA algorithm.
}
