% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{revCumSum}
\alias{revCumSum}
\title{Calculate Reverse Cumulative Sums of Vector Elements}
\usage{
revCumSum(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
A vector containing the reverse cumulative sums of the input vector.
}
\description{
This function calculates the reverse cumulative sums of elements in a given vector.
}
