% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Round}
\alias{Round}
\title{Round Vector Elements}
\usage{
Round(v, eps)
}
\arguments{
\item{v}{A vector.}

\item{eps}{A small number representing the rounding threshold.}
}
\value{
A vector with rounded elements.
}
\description{
This function rounds the elements of a given vector to the nearest integer, based on a specified threshold.
}
