% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roundvec}
\alias{roundvec}
\title{Round Close Values to the Same}
\usage{
roundvec(beta, eps)
}
\arguments{
\item{beta}{The vector of coefficients.}

\item{eps}{The tolerance for rounding.}
}
\description{
This function rounds close values in a vector to the same value based on a given tolerance.
}
