% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{RowSums}
\alias{RowSums}
\title{Calculating the row sums}
\usage{
RowSums(x, ...)
}
\arguments{
\item{x}{The input, which can be a number, a vector, or a matrix.}

\item{...}{see \code{\link[base:colSums]{base::rowSums()}} for details}
}
\value{
A vector of sums.
}
\description{
This function calculates the row sums of a given input, which can be a number, a vector, or a matrix. Before performing the calculation, the function checks whether the input is a vector or a matrix. If it's a vector, the function treats it as a single column matrix and does nothing. If it's a matrix, the function calculates the row sums.
}
\examples{
RowSums(array(1:60, c(3,4,5)), dims = 1L)
# Output: 590 610 630

}
