% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{runOpera}
\alias{runOpera}
\title{Performing cancer staging}
\usage{
runOpera(
  ncat,
  dat,
  TimeN,
  yN,
  cenN,
  covN = NULL,
  withCov = F,
  filepath = "./operawp_results/",
  riskVariables,
  riskFactors,
  riskNames,
  useLassoT = F,
  useOPERA = F,
  usePruning = F,
  edges = NULL,
  ifE = F,
  type = "surv",
  getsAll = F,
  useAIC = F,
  useIbs = F,
  useLRT = T,
  prefix_stage = NULL,
  threshold = NULL,
  coarse_pruning = F,
  fine_pruning = F,
  fine_pruning_quad = F,
  eps = 10^-4,
  eps_lasso = 10^-2,
  maxiter = 10,
  maxiter_lasso = 200,
  GIC = NULL,
  useBIC = F,
  perc = 0.1,
  minObs = 30,
  plt = T,
  pic_width = 8,
  pic_height = 8,
  legend_size = 0.7,
  x_pos = -1,
  yratio = 0.1,
  yaxis_min = 0,
  xpos_bs = 0.02,
  x_label = "Time (Days)",
  ...
)
}
\arguments{
\item{ncat}{A vector specifying the number of categories in each variable or risk factor.}

\item{dat}{The dataset}

\item{TimeN}{The variable name for survival times}

\item{yN}{The variable name for a binary outcome}

\item{cenN}{The variable name for censoring times or a binary outcome}

\item{covN}{The variable names for covariates}

\item{withCov}{Whether any covariates need adjusting}

\item{filepath}{The file path to save all the results}

\item{riskVariables}{The variable names for all risk factors}

\item{riskFactors}{The list of names for all levels of all risk factors.}

\item{riskNames}{The names of all risk factors.}

\item{useLassoT}{Whether lasso tree method is used}

\item{useOPERA}{Whether OPERA is used}

\item{usePruning}{Whether pruning is used}

\item{edges}{edges needed when a risk factor lacks total orderign}

\item{ifE}{whether edges are given}

\item{type}{The outcome type, either survival outcome or binary outcome.}

\item{getsAll}{If all criteria in pruning are used}

\item{useAIC}{If AIC is used as a criterion in pruning}

\item{useIbs}{If IBS is used as a criterion in pruning}

\item{useLRT}{If likelihood ratio test is used as a criterion in pruning}

\item{prefix_stage}{The number of stages needed.}

\item{threshold}{The type-I error rate for the likelihood ratio test.}

\item{coarse_pruning}{A boolean indicating whether coarse pruning is applied.}

\item{fine_pruning}{A boolean indicating whether fine pruning using exhaustive search is applied.}

\item{fine_pruning_quad}{A boolean indicating whether fine pruning using quadratic programming constraint is applied.}

\item{eps}{The tolerance level except for lasso tree method}

\item{eps_lasso}{The tolerance level for lasso tree method}

\item{maxiter}{The maximum number of iterations in each approximation step except for lasso tree method}

\item{maxiter_lasso}{The maximum number of iterations in each approximation step for lasso tree method}

\item{GIC}{The penalty term used in the criterion except for lasso tree method.}

\item{useBIC}{Whether BIC is used for lasso tree method. If false, AIC is used.}

\item{perc}{The minimum proportion of patients required in each stage}

\item{minObs}{The minimum number of patients required in each stage}

\item{plt}{Whether results are needed to be plotted}

\item{pic_width}{The width of the figure of the network of risk categories}

\item{pic_height}{The height of the figure of the network of risk categories}

\item{legend_size}{The size of the legend in the figure of the network of risk categories}

\item{x_pos}{The horizontal position of the legend in the figure of the network of risk categories}

\item{yratio}{The ratio proportional to the number of levels in each factor to control the vertical position of the legend in the figure of the network of risk categories}

\item{yaxis_min}{The minimum value of the y-axis for the Kaplan-Meier curves}

\item{xpos_bs}{The horizontal position of the label for the Brier score}

\item{x_label}{The label of the x-axis for the Kaplan-Meier curves}
}
\value{
The staging result
}
\description{
This function performs cancer staging
}
