% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{sbrierC}
\alias{sbrierC}
\title{Calculating the integrated Brier score}
\usage{
sbrierC(obj, pred)
}
\arguments{
\item{obj}{An object of class \code{Surv} representing the survival data.}

\item{pred}{A list of \code{survfit} objects representing the predicted survival probabilities.}
}
\value{
The integrated Brier score and the starting and ending time points.
}
\description{
The function computes the integrated Brier score, which is a measure of the accuracy of survival predictions. It assesses the agreement between the observed survival outcomes and the predicted survival probabilities over a range of time points. The integrated Brier score provides an overall evaluation of the predictive performance.
}
\references{
Graf, E., Schmoor, C., Sauerbrei, W., & Schumacher, M. (1999). Assessment and comparison of prognostic classification schemes for survival data. Statistics in Medicine, 18(17-18), 2529-2545. Retrieved from \href{https://rdrr.io/cran/ipred/src/R/sbrier.R}{here}
}
