% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sum}
\alias{Sum}
\title{Calculate Sum of Vector Elements}
\usage{
Sum(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
The sum of the vector elements.
}
\description{
This function calculates the sum of elements in a given vector.
}
