% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_risk}
\alias{sum_risk}
\title{Perform Special Summation Over Risk Sets}
\usage{
sum_risk(indset, value_vec, ind_vec, p)
}
\arguments{
\item{indset}{The index matrix of the risk set.}

\item{value_vec}{The values to be summed over. In our problem, it represents the exponential of eta.}

\item{ind_vec}{The vector indicating the failure times to be summed over.}

\item{p}{The power number.}
}
\description{
This function calculates a special summation over risk sets based on given parameters.
}
