% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SumL}
\alias{SumL}
\title{Calculating the sum of a vector}
\usage{
SumL(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
The sum of the elements in the vector.
}
\description{
This function calculates the sum of a given vector.
}
