% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{val_pen}
\alias{val_pen}
\title{Calculate Value of Penalty Term}
\usage{
val_pen(Apo, w, nb, beta)
}
\arguments{
\item{Apo}{The partial order matrix.}

\item{w}{The adaptive parameter.}

\item{nb}{The number of boundaries}

\item{beta}{The vector of coefficients.}
}
\value{
The value of the penalty term.
}
\description{
This function calculates the value of the penalty term based on the given parameters.
}
