% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wInf}
\alias{wInf}
\title{Check for Inf or NaN in Vector}
\usage{
wInf(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
A boolean value indicating whether the vector contains Inf or NaN.
}
\description{
This function checks whether a given vector contains any elements that are either Inf (infinity) or NaN (not-a-number).
}
