% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPERA.R
\name{Zmatrix}
\alias{Zmatrix}
\title{Generating the design matrix for risk categories}
\usage{
Zmatrix(dat, colNames)
}
\arguments{
\item{dat}{The dataset that includes a \code{variable} column specifying each observation's node.}

\item{colNames}{The node names.}
}
\value{
A matrix \code{Z} representing the node to which each observation resides.
}
\description{
The function generates a matrix \code{Z} that indicates the risk categories or nodes to which each observation in the dataset belongs. Each row of the matrix corresponds to an observation, and each column represents a node. The value in each cell of the matrix indicates whether the observation belongs to the corresponding node (1 if yes, 0 if no).
}
